+++ Data/SBV/Control/Utils.hs
@@ -131,12 +131,11 @@ syncUpSolver :: Bool -> IncState -> Query ()
 syncUpSolver afterAPush is = do
         cfg <- getConfig
         ls  <- io $ do let swap  (a, b)        = (b, a)
-                           swapc ((_, a), b)   = (b, a)
                            cmp   (a, _) (b, _) = a `compare` b
                            arrange (i, (at, rt, es)) = ((i, at, rt), es)
                        inps  <- reverse <$> readIORef (rNewInps is)
                        ks    <- readIORef (rNewKinds is)
-                       cnsts <- sortBy cmp . map swapc . Map.toList <$> readIORef (rNewConsts is)
+                       cnsts <- sortBy cmp . map swap . Map.toList <$> readIORef (rNewConsts is)
                        arrs  <- IMap.toAscList <$> readIORef (rNewArrs is)
                        tbls  <- map arrange . sortBy cmp . map swap . Map.toList <$> readIORef (rNewTbls is)
                        uis   <- Map.toAscList <$> readIORef (rNewUIs is)